seed_project/requirement.txt
torch
torchvision
matplotlib
seaborn
scikit-learn
opencv-python
grad-cam
Pillow
numpy
